Graph view snapshot has been overhauled and improved
Previously, the taken snapshot only showed current viewport, instead of
the whole graph. Also, the resolution depended on the physical
dimensions of the viewport. Now, the snapshot includes the visible graph
elements, even if they are not in the viewport, with a more than decent
resolution.

As the cytoscape-svg plugin has also been integrated, now it is possible
to save the snapshot in SVG (vector) format.

Last, migration to latest version of fontawesome.